# DOCS/api-swagger.json
{
   "openapi":"3.0.1",
   "info":{
      "title":"OpenAPI definition",
      "version":"v0"
   },
   "servers":[
      {
         "url":"http://localhost:8080",
         "description":"Generated server url"
      }
   ],
   "paths":{
      "/pessoa/criar":{
         "post":{
            "tags":[
               "API para gerenciamento de pessoas"
            ],
            "summary":"Cria um novo registro de pessoa",
            "operationId":"criarPessoa",
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/PessoaRequest"
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ResponsePessoa"
                        }
                     }
                  }
               }
            }
         }
      },
      "/conta/sacar":{
         "post":{
            "tags":[
               "API para gerenciamento de contas"
            ],
            "summary":"Faz um saque em uma conta",
            "operationId":"sacarValor",
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/SaqueRequest"
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ResponseTransacao"
                        }
                     }
                  }
               }
            }
         }
      },
      "/conta/depositar":{
         "post":{
            "tags":[
               "API para gerenciamento de contas"
            ],
            "summary":"Faz um deposito em uma conta",
            "operationId":"depositarValor",
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/DepositoRequest"
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ResponseTransacao"
                        }
                     }
                  }
               }
            }
         }
      },
      "/conta/criar":{
         "post":{
            "tags":[
               "API para gerenciamento de contas"
            ],
            "summary":"Cria uma nova conta",
            "operationId":"criarConta",
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/ContaRequest"
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ResponseConta"
                        }
                     }
                  }
               }
            }
         }
      },
      "/conta/bloquear/{idConta}":{
         "post":{
            "tags":[
               "API para gerenciamento de contas"
            ],
            "summary":"Bloqueia uma conta",
            "operationId":"bloquearConta",
            "parameters":[
               {
                  "name":"idConta",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"integer",
                     "format":"int32"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ResponseConta"
                        }
                     }
                  }
               }
            }
         }
      },
      "/conta/saldo/{idConta}":{
         "get":{
            "tags":[
               "API para gerenciamento de contas"
            ],
            "summary":"Obtem o saldo de uma conta",
            "operationId":"obterSaldo",
            "parameters":[
               {
                  "name":"idConta",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"integer",
                     "format":"int32"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ResponseBigDecimal"
                        }
                     }
                  }
               }
            }
         }
      },
      "/conta/extrato":{
         "get":{
            "tags":[
               "API para gerenciamento de contas"
            ],
            "summary":"Retorna o extrato de transacoes de uma conta",
            "operationId":"extratoConta",
            "parameters":[
               {
                  "name":"idConta",
                  "in":"query",
                  "required":true,
                  "schema":{
                     "type":"integer",
                     "format":"int32"
                  }
               },
               {
                  "name":"dataInicial",
                  "in":"query",
                  "required":false,
                  "schema":{
                     "type":"string"
                  }
               },
               {
                  "name":"dataFinal",
                  "in":"query",
                  "required":false,
                  "schema":{
                     "type":"string"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "content":{
                     "*/*":{
                        "schema":{
                           "$ref":"#/components/schemas/ResponseListTransacao"
                        }
                     }
                  }
               }
            }
         }
      }
   },
   "components":{
      "schemas":{
         "PessoaRequest":{
            "required":[
               "cpf",
               "dataNascimento",
               "nome"
            ],
            "type":"object",
            "properties":{
               "nome":{
                  "type":"string",
                  "description":"Nome da pessoa",
                  "example":"Everton M"
               },
               "cpf":{
                  "pattern":"[0-9]",
                  "type":"string",
                  "description":"CPF da pessoa, somente numeros",
                  "example":"12345678900"
               },
               "dataNascimento":{
                  "pattern":"^(3[01]|[12][0-9]|0[1-9])/(1[0-2]|0[1-9])/[0-9]{4}$",
                  "type":"string",
                  "description":"Data de nascimento no formato DD/MM/AAAA",
                  "example":"22/09/1987"
               }
            }
         },
         "Pessoa":{
            "type":"object",
            "properties":{
               "idPessoa":{
                  "type":"integer",
                  "format":"int32"
               },
               "nome":{
                  "type":"string"
               },
               "cpf":{
                  "type":"string"
               },
               "dataNascimento":{
                  "type":"string",
                  "format":"date-time"
               }
            }
         },
         "ResponsePessoa":{
            "type":"object",
            "properties":{
               "mensagem":{
                  "type":"string"
               },
               "payload":{
                  "$ref":"#/components/schemas/Pessoa"
               }
            }
         },
         "SaqueRequest":{
            "required":[
               "idConta",
               "valor"
            ],
            "type":"object",
            "properties":{
               "idConta":{
                  "type":"integer",
                  "description":"Codigo da pessoa",
                  "format":"int32",
                  "example":3
               },
               "valor":{
                  "type":"number",
                  "description":"Valor do saque",
                  "example":15.0
               }
            }
         },
         "ResponseTransacao":{
            "type":"object",
            "properties":{
               "mensagem":{
                  "type":"string"
               },
               "payload":{
                  "$ref":"#/components/schemas/Transacao"
               }
            }
         },
         "Transacao":{
            "type":"object",
            "properties":{
               "idTransacao":{
                  "type":"integer",
                  "format":"int32"
               },
               "idConta":{
                  "type":"integer",
                  "format":"int32"
               },
               "valor":{
                  "type":"number"
               },
               "dataTransacao":{
                  "type":"string",
                  "format":"date-time"
               },
               "tipoTransacao":{
                  "type":"integer",
                  "format":"int32"
               }
            }
         },
         "DepositoRequest":{
            "required":[
               "idConta",
               "valor"
            ],
            "type":"object",
            "properties":{
               "idConta":{
                  "type":"integer",
                  "description":"Codigo da conta",
                  "format":"int32",
                  "example":3
               },
               "valor":{
                  "type":"number",
                  "description":"Valor do deposito",
                  "example":50.9
               }
            }
         },
         "ContaRequest":{
            "required":[
               "idPessoa",
               "limiteSaqueDiario",
               "tipoConta"
            ],
            "type":"object",
            "properties":{
               "idPessoa":{
                  "type":"integer",
                  "description":"Codigo da pessoa",
                  "format":"int32",
                  "example":3
               },
               "limiteSaqueDiario":{
                  "type":"number",
                  "description":"Limite diario de saque. Duas casas decimais separados ponto",
                  "example":1500.99
               },
               "tipoConta":{
                  "type":"string",
                  "description":"Tipo da conta. Pode ser CORRENTE ou POUPANCA",
                  "example":"CORRENTE",
                  "enum":[
                     "CORRENTE",
                     "POUPANCA",
                     "CORRENTE",
                     "POUPANCA"
                  ]
               }
            }
         },
         "Conta":{
            "type":"object",
            "properties":{
               "idConta":{
                  "type":"integer",
                  "format":"int32"
               },
               "idPessoa":{
                  "type":"integer",
                  "format":"int32"
               },
               "saldo":{
                  "type":"number"
               },
               "limiteSaqueDiario":{
                  "type":"number"
               },
               "flagAtivo":{
                  "type":"boolean"
               },
               "tipoConta":{
                  "type":"integer",
                  "format":"int32"
               },
               "dataCriacao":{
                  "type":"string",
                  "format":"date-time"
               }
            }
         },
         "ResponseConta":{
            "type":"object",
            "properties":{
               "mensagem":{
                  "type":"string"
               },
               "payload":{
                  "$ref":"#/components/schemas/Conta"
               }
            }
         },
         "ResponseBigDecimal":{
            "type":"object",
            "properties":{
               "mensagem":{
                  "type":"string"
               },
               "payload":{
                  "type":"number"
               }
            }
         },
         "ResponseListTransacao":{
            "type":"object",
            "properties":{
               "mensagem":{
                  "type":"string"
               },
               "payload":{
                  "type":"array",
                  "items":{
                     "$ref":"#/components/schemas/Transacao"
                  }
               }
            }
         }
      }
   }
}